+++ src/query/helper.rs
@@ -731,12 +731,11 @@ pub trait QueryFilter: Sized {
     /// Like above, but using the `ANY` operator. Postgres only.
     ///
     /// ```
-    /// use sea_orm::sea_query::{Expr, ExprTrait, extension::postgres::PgFunc};
     /// use sea_orm::{DbBackend, entity::*, query::*, tests_cfg::cake};
     ///
     /// assert_eq!(
     ///     cake::Entity::find()
-    ///         .filter(Expr::col((cake::Entity, cake::Column::Id)).eq(PgFunc::any(vec![4, 5])))
+    ///         .filter(cake::Column::Id.eq_any(vec![4, 5]))
     ///         .build(DbBackend::Postgres)
     ///         .to_string(),
     ///     r#"SELECT "cake"."id", "cake"."name" FROM "cake" WHERE "cake"."id" = ANY(ARRAY [4,5])"#